NotifierTest: Do not depend on the time zone
Date.toString() includes the time zone, which by default is the timezone
of the host running the JVM. So time zones may differ depending on the
host and we must not rely on a specific time zone to be part of expected
test output. Fix this by using the host's time zone as part of the
expected output.